requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.21.0
tqdm>=4.65.0
rich>=13.0.0
python-dotenv>=1.0.0
safetensors>=0.3.0
sentencepiece>=0.1.99
protobuf>=3.20.0
wandb>=0.15.0